change the method of calling search_fragment. now it is called by the OnItemClickListener method from CountriesDetailsFragment directly
