+++ presentation/src/main/java/com/example/presentation/countries/CountriesListFragment.kt
@@ -4,13 +4,16 @@ import android.os.Bundle
 import android.view.View
 import android.widget.Toast
 import androidx.activity.addCallback
+import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.setFragmentResult
 import androidx.recyclerview.widget.LinearLayoutManager
 import by.kirich1409.viewbindingdelegate.viewBinding
 import com.example.presentation.MainFragment
 import com.example.presentation.R
 import com.example.presentation.databinding.CountriesListFragmentBinding
 import com.google.android.material.divider.MaterialDividerItemDecoration
+import com.google.gson.Gson
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -55,18 +58,35 @@ class CountriesListFragment : Fragment(R.layout.countries_list_fragment) {
                     country.name?.common
                 }
                 withContext(Dispatchers.Main) {
-
+                    val countriesListAdapter = countriesList?.let { CountriesAdapter(it) }
+                    binding.countriesList.adapter = countriesListAdapter
                     binding.countriesList.layoutManager = LinearLayoutManager(requireContext())
-                    binding.countriesList.adapter = countriesList?.let { CountriesAdapter(it) }
                     binding.countriesList.addItemDecoration(divider)
+                    countriesListAdapter?.setOnItemClickListener(object : CountriesAdapter.OnItemClickListener {
+                        override fun onItemClick(position: Int) {
+                            val country = countriesList[position]
+                            val jsonString = Gson().toJson(country)
+                            setFragmentResult(
+                                EXTRA_COUNTRY_REQUESTED_KEY,
+                                bundleOf(COUNTRY_BUNDLE_KEY to jsonString)
+                            )
 
+                            requireActivity().supportFragmentManager.beginTransaction()
+                                .replace(R.id.FragmentContainerView, CountriesDetailsFragment()).commit()
+                            }
+                        })
                 }
             }
             else{
                 countriesListResponse.message().also{
                     Toast.makeText(requireContext(), it, Toast.LENGTH_LONG).show()
                 }
+                
             }
         }
+    }
+    companion object {
+        private const val EXTRA_COUNTRY_REQUESTED_KEY = "EXTRA_COUNTRY_REQUESTED_KEY"
+        private const val COUNTRY_BUNDLE_KEY = "COUNTRY_BUNDLE_KEY"
     }
 }
