+++ presentation/src/main/res/layout/fragment_main.xml
@@ -30,16 +30,4 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/arts_list_button"/>
 
-    <Button
-        android:id="@+id/search_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/search_countries_button"
-        style="@style/Widget.AppCompat.Button.Borderless"
-        android:background="@drawable/button_background"
-        android:layout_marginTop="5dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/countries_list_button"/>
-
 </androidx.constraintlayout.widget.ConstraintLayout>
